+++ next.config.mjs
@@ -1,3 +1,5 @@
+import { withSentryConfig } from '@sentry/nextjs/config';
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
@@ -25,5 +27,14 @@ const nextConfig = {
   },
 };
 
-export default nextConfig;
-
+export default withSentryConfig(nextConfig, {
+  org: 'makelogoai-app',
+  project: 'pollenmonitor',
+  silent: !process.env.CI,
+  widenClientFileUpload: true,
+  webpack: {
+    treeshake: {
+      removeDebugLogging: true,
+    },
+  },
+});